add goto
