+++ src/stage/commands/AddSome.js
@@ -4,20 +4,11 @@ const Command = require("../../main/Command");
 
 module.exports = class AddSome extends Command {
 
-    constructor(stage, data) {
-        console.log("AddSome Constructor");
-        console.log(stage);
-        console.log(data);
-        super('add-some', stage, data);
+    constructor(data) {
+        super('AddSome', data);
     }
 
-    async doAction() {
-        await console.log("AddSome doAction")
-        console.log(this.data);
-        let store = this.stage.store;
-        if(typeof store !== 'undefined'){
-            store.counter += this.data;
-        }
-        return "ok"
+    async do(store){
+        store.counter = Number(store.counter) + Number(this.data.amt);
     }
 };
