+++ src/stage/commands/GameTick.js
@@ -8,7 +8,7 @@ module.exports = class GameTick extends Command {
         super('GameTick', data);
     }
 
-    async do(state){
+    async do(store){
         return 'ok'
     }
 };
