+++ src/stage/commands/SelectCard.js
@@ -8,8 +8,7 @@ module.exports = class SelectCard extends Command {
         super('SelectCard', data);
     }
 
-    async do(state){
-        let store = this.stage.store;
+    async do(store){
         let data = this.data;
         if(typeof data !== 'undefined') {
             let actor = store.actors[data.actorId];
